requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117

hydra-core==1.2.0
kaggle==1.5.12
libzero==0.0.8
matplotlib==3.5.2
numpy==1.23.1
rtdl==0.0.13
scikit-learn==1.1.2
tensorboard==2.9.1
timm==0.6.7
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
xgboost==1.6.1
catboost==1.1.1
pytorch_pfn_extras==0.5.8
imbalanced-learn==0.9.1
GitPython==3.1.27
pandas==1.5.0
einops==0.5.0
optuna==3.0.3
transtab==0.0.2
PyMySQL==1.0.2
cryptography==39.0.1
sentence-transformers==2.2.2
bitsandbytes==0.37.1
accelerate==0.17.1
